+++ jena-weather/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tfjs-examples-addition-rnn",
+  "name": "tfjs-examples-jena-weather",
   "version": "0.1.0",
   "description": "",
   "main": "index.js",
