+++ jena-weather/ui.js
@@ -15,16 +15,21 @@
  * =============================================================================
  */
 
+import {plotData} from './index';
+
 const statusElement = document.getElementById('status');
-
-export function logStatus(message) {
-  statusElement.innerText = message;
-}
-
 const timeSpanSelect = document.getElementById('time-span');
 const selectSeries1 = document.getElementById('data-series-1');
 const selectSeries2 = document.getElementById('data-series-2');
 const dataNormalizedCheckbox = document.getElementById('data-normalized');
+const dateTimeRangeSpan = document.getElementById('date-time-range');
+const dataPrevButton = document.getElementById('data-prev');
+const dataNextButton = document.getElementById('data-next');
+const dataScatterCheckbox = document.getElementById('data-scatter');
+
+export function logStatus(message) {
+  statusElement.innerText = message;
+}
 
 export function populateSelects(dataObj) {
   const columnNames = ['None'].concat(dataObj.getDataColumnNames());
@@ -49,4 +54,75 @@ export function populateSelects(dataObj) {
   }
   timeSpanSelect.value = 'week';
   dataNormalizedCheckbox.checked = true;
+}
+
+export const TIME_SPAN_RANGE_MAP = {
+  hour: 6,
+  day: 6 * 24,
+  week: 6 * 24 * 7,
+  tenDays: 6 * 24 * 10,
+  month: 6 * 24 * 30,
+  year: 6 * 24 * 365,
+  full: null
+};
+
+export const TIME_SPAN_STRIDE_MAP = {
+  day: 1,
+  week: 1,
+  tenDays: 6,
+  month: 6,
+  year: 6 * 6,
+  full: 6 * 24
+};
+
+export let currBeginIndex = 0;
+
+export function updateDateTimeRangeSpan(jenaWeatherData) {
+  const timeSpan = timeSpanSelect.value;
+  const currEndIndex = currBeginIndex + TIME_SPAN_RANGE_MAP[timeSpan];
+  const begin =
+      new Date(jenaWeatherData.getTime(currBeginIndex)).toLocaleDateString();
+  const end =
+      new Date(jenaWeatherData.getTime(currEndIndex)).toLocaleDateString();
+  dateTimeRangeSpan.textContent = `${begin} - ${end}`;
+}
+
+export function updateScatterCheckbox() {
+  const series1 = selectSeries1.value;
+  const series2 = selectSeries2.value;
+  dataScatterCheckbox.disabled = series1 === 'None' || series2 === 'None';
+}
+
+dataPrevButton.addEventListener('click', () => {
+  const timeSpan = timeSpanSelect.value;
+  currBeginIndex -= Math.round(TIME_SPAN_RANGE_MAP[timeSpan] / 8);
+  if (currBeginIndex >= 0) {
+    plotData();
+  } else {
+    currBeginIndex = 0;
+  }
+});
+
+dataNextButton.addEventListener('click', () => {
+  const timeSpan = timeSpanSelect.value;
+  currBeginIndex += Math.round(TIME_SPAN_RANGE_MAP[timeSpan] / 8);
+  plotData();
+});
+
+timeSpanSelect.addEventListener('change', () => {
+  plotData();
+});
+selectSeries1.addEventListener('change', plotData);
+selectSeries2.addEventListener('change', plotData);
+dataNormalizedCheckbox.addEventListener('change', plotData);
+dataScatterCheckbox.addEventListener('change', plotData);
+
+export function getDataVizOptions() {
+  return {
+    timeSpan: timeSpanSelect.value,
+    series1: selectSeries1.value,
+    series2: selectSeries2.value,
+    normalize: dataNormalizedCheckbox.checked,
+    scatter: dataScatterCheckbox.checked
+  };
 }
